Move TestReconcileDeprecatedGlobalConfig to proper file
